Add a new type of function: constructor_variable to hold state variable initialization Add Function.FunctionType to determine if the function is a constructor/fallback/constructor_variables API change: function.is_fallback() becomes a property
