+++ slither/solc_parsing/declarations/contract.py
@@ -1,7 +1,10 @@
 import logging
 
 from slither.core.declarations.contract import Contract
+from slither.core.declarations.function import Function, FunctionType
 from slither.core.declarations.enum import Enum
+from slither.core.cfg.node import Node, NodeType
+from slither.core.expressions import AssignmentOperation, Identifier, AssignmentOperationType
 from slither.slithir.variables import StateIRVariable
 from slither.solc_parsing.declarations.event import EventSolc
 from slither.solc_parsing.declarations.function import FunctionSolc
@@ -319,7 +322,6 @@ class ContractSolc04(Contract):
         :return:
         """
         all_elements = {}
-        accessible_elements = {}
 
         for father in self.inheritance:
             for element in getter(father):
@@ -367,6 +369,49 @@ class ContractSolc04(Contract):
                 except VariableNotFound:
                     pass
         return
+
+
+    def _create_node(self, func, counter, variable):
+        # Function uses to create node for state variable declaration statements
+        node = Node(NodeType.STANDALONE, counter)
+        node.set_offset(variable.source_mapping, self.slither)
+        node.set_function(func)
+        func.add_node(node)
+        print(variable.expression)
+        expression = AssignmentOperation(Identifier(variable),
+                                                    variable.expression,
+                                                    AssignmentOperationType.ASSIGN,
+                                                    variable.type)
+
+        node.add_expression(expression)
+        return node
+
+    def add_constructor_variables(self):
+        if self.state_variables:
+            found_candidate = False
+            for (idx, variable_candidate) in enumerate(self.state_variables):
+                if variable_candidate.expression and not variable_candidate.is_constant:
+                    found_candidate = True
+                    break
+            if found_candidate:
+                constructor_variable = Function()
+                constructor_variable.set_function_type(FunctionType.CONSTRUCTOR_VARIABLES)
+                constructor_variable.set_contract(self)
+                constructor_variable.set_contract_declarer(self)
+                constructor_variable.set_visibility('internal')
+                self._functions[constructor_variable.canonical_name] = constructor_variable
+
+                prev_node = self._create_node(constructor_variable, 0, variable_candidate)
+                counter = 1
+                for v in self.state_variables[idx+1:]:
+                    if v.expression and not v.is_constant:
+                        next_node = self._create_node(constructor_variable, counter, v)
+                        prev_node.add_son(next_node)
+                        next_node.add_father(prev_node)
+                        counter += 1
+
+
+
 
     def analyze_state_variables(self):
         for var in self.variables:
